+++ lib/hexapdf/importer.rb
@@ -71,14 +71,18 @@ module HexaPDF
     # Imports the given +object+ (belonging to the +source+ document) by completely copying it and
     # all referenced objects into the +destination+ object.
     #
+    # If the +allow_all+ argument is set to +true+, then the usually omitted catalog and page tree
+    # node objects (see the class description for details) are also copied which allows one to make
+    # an in-memory duplicate of a HexaPDF::Document object.
+    #
     # Specifying +source+ is optionial if it can be determined through +object+.
     #
     # After the operation is finished, all state is discarded. This means that another call to this
     # method for the same object will yield a new - and different - object. This is in contrast to
     # using ::for together with #import which remembers and returns already imported objects (which
     # is generally what one wants).
-    def self.copy(destination, object, source: nil)
-      new(NullableWeakRef.new(destination)).import(object, source: source)
+    def self.copy(destination, object, allow_all: false, source: nil)
+      new(NullableWeakRef.new(destination), allow_all: allow_all).import(object, source: source)
     end
 
     private_class_method :new
@@ -86,9 +90,10 @@ module HexaPDF
     attr_reader :destination #:nodoc:
 
     # Initializes a new importer that can import objects to the +destination+ document.
-    def initialize(destination)
+    def initialize(destination, allow_all: false)
       @destination = destination
       @mapper = {}
+      @allow_all = allow_all
     end
 
     SourceWrapper = Struct.new(:source) #:nodoc:
@@ -136,7 +141,7 @@ module HexaPDF
         internal_import(wrapper.source.object(object), wrapper)
       when HexaPDF::Object
         wrapper.source ||= object.document
-        if object.type == :Catalog || object.type == :Pages
+        if !@allow_all && (object.type == :Catalog || object.type == :Pages)
           @mapper[object.data] = nil
         elsif (mapped_object = @mapper[object.data]&.__getobj__) && !mapped_object.null?
           mapped_object